requirements.txt
allure-pytest==2.13.2
playwright==1.53.0
pytest==7.4.3
pytest-html==4.1.1
pytest-playwright==0.4.3
pytest-xdist==3.5.0
python-dotenv==1.0.0
colorlog==6.8.0
